Modify Degrees to Int

+++ app/src/main/java/com/example/weatherapp/data/models/WindModel.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class WindModel(
     @SerializedName("speed") val speed: Float,
-    @SerializedName("deg") val deg: Float,
+    @SerializedName("deg") val deg: Int,
     @SerializedName("gust") val gust: Float,
 )
